requirements.txt
numpy==1.26.2
pyobjc-core==10.1
pyobjc-framework-Cocoa==10.1
pyobjc-framework-Metal==10.1
